Script extraxxion_datos: Implementación de validación en caso de que el archivo CSV ya existe.

+++ EC04/VISUALIZACION DE INFRAESTRUCTURA DE REDES PUBLICAS EN EL DISTRITO METROPOLITANO DE QUITO/manejo_de_datos/extraccion_datos.py
@@ -40,6 +40,10 @@ def guardar_datos(df, directorio_salida):
     procesar para uso posterior (archivo CSV).
     """
     if df is not None:
+        if os.path.exists(directorio_salida):
+            print(f"Advertencia: El archivo ya existe en {directorio_salida}.")
+            print("El archivo no se sobreescribirá.") 
+            return False
         try:
             df.to_csv(directorio_salida, index=False)
             print(f"Datos sin procesar guardados en: {directorio_salida}\n")
